# docs/swagger.yaml
basePath: /
definitions:
  model.APIError:
    properties:
      message:
        type: string
    type: object
  model.Delivery:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    required:
    - address
    - city
    - email
    - name
    - phone
    - region
    - zip
    type: object
  model.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.APIError'
    type: object
  model.Item:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: integer
      rid:
        type: string
      sale:
        type: integer
      size:
        type: string
      status:
        type: integer
      total_price:
        type: integer
      track_number:
        type: string
    required:
    - brand
    - chrt_id
    - name
    - nm_id
    - price
    - rid
    - sale
    - size
    - status
    - total_price
    - track_number
    type: object
  model.Order:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/model.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/model.Payment'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    required:
    - customer_id
    - date_created
    - delivery
    - delivery_service
    - entry
    - items
    - locale
    - oof_shard
    - order_uid
    - payment
    - shardkey
    - sm_id
    - track_number
    type: object
  model.Payment:
    properties:
      amount:
        type: integer
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        type: integer
      delivery_cost:
        type: integer
      goods_total:
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    required:
    - amount
    - bank
    - currency
    - custom_fee
    - delivery_cost
    - goods_total
    - payment_dt
    - provider
    - transaction
    type: object
host: localhost:8081
info:
  contact: {}
  title: Orders
  version: "1.0"
paths:
  /order/{id}:
    get:
      description: Get order by order_uid
      operationId: get-order
      parameters:
      - description: Order uid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Invalid uid format
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "504":
          description: Server timeout
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get order
      tags:
      - orders
swagger: "2.0"